feat(irl-event): email notification code modified
